storage: just move spdk to a separate file
