+++ storage/server.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"encoding/json"
-	"io/ioutil"
 	"context"
 	"flag"
 	"fmt"
@@ -23,32 +22,6 @@ type server struct {
 	pb.UnimplementedNVMeControllerServiceServer
 	pb.UnimplementedNVMeNamespaceServiceServer
 	pb.UnimplementedNVMfRemoteControllerServiceServer
-}
-
-func spdkCommunicate(buf []byte) ([]byte, error) {
-	// TODO: use rpc_sock variable
-	conn, err := net.Dial("unix", *rpc_sock)
-	if err != nil {
-		log.Fatal(err)
-	}
-	_, err = conn.Write(buf)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = conn.(*net.UnixConn).CloseWrite()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	reply, err := ioutil.ReadAll(conn)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Println(string(reply))
-
-	return reply, err
 }
 
 func (s *server) NVMeSubsystemCreate(ctx context.Context, in *pb.NVMeSubsystemCreateRequest) (*pb.NVMeSubsystemCreateResponse, error) {
